Added the unit tests for the getNetDevicesFromPci function
